Update curl usage and ReadFromSocket.

+++ sdk/core/azure-core/src/http/curl/curl_connection_private.hpp
@@ -80,7 +80,7 @@ namespace Azure { namespace Core { namespace Http {
      * there is no more data to get from the socket.
      *
      */
-    virtual int64_t ReadFromSocket(uint8_t* buffer, int64_t bufferSize, Context const& context) = 0;
+    virtual size_t ReadFromSocket(uint8_t* buffer, size_t bufferSize, Context const& context) = 0;
 
     /**
      * @brief This method will use libcurl socket to write all the bytes from buffer.
@@ -184,7 +184,7 @@ namespace Azure { namespace Core { namespace Http {
        * @return return the numbers of bytes pulled from socket. It can be less than what it was
        * requested.
        */
-      int64_t ReadFromSocket(uint8_t* buffer, int64_t bufferSize, Context const& context) override;
+      size_t ReadFromSocket(uint8_t* buffer, size_t bufferSize, Context const& context) override;
 
       /**
        * @brief This method will use libcurl socket to write all the bytes from buffer.
